Extracted isIAPSupported to a separate class
