+++ libs/iap/src/main/java/com/woocommerce/android/iap/internal/planpurchase/IAPPurchaseWPComPlanActionsImpl.kt
@@ -9,7 +9,6 @@ import com.woocommerce.android.iap.internal.model.IAPProductDetailsResponse
 import com.woocommerce.android.iap.internal.model.IAPPurchaseResult
 import com.woocommerce.android.iap.pub.IAPActivityWrapper
 import com.woocommerce.android.iap.pub.PurchaseWPComPlanActions
-import com.woocommerce.android.iap.pub.model.IAPSupportedResult
 import com.woocommerce.android.iap.pub.model.WPComIsPurchasedResult
 import com.woocommerce.android.iap.pub.model.WPComPlanProduct
 import com.woocommerce.android.iap.pub.model.WPComProductResult
@@ -17,8 +16,6 @@ import com.woocommerce.android.iap.pub.model.WPComPurchaseResult
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.merge
-
-private const val SUPPORTED_CURRENCY = "USD"
 
 internal class IAPPurchaseWPComPlanActionsImpl(
     private val purchaseWpComPlanHandler: IAPPurchaseWpComPlanHandler,
@@ -69,17 +66,7 @@ internal class IAPPurchaseWPComPlanActionsImpl(
         }
     }
 
-    override suspend fun isIAPSupported(): IAPSupportedResult {
-        return when (val response = iapManager.fetchIAPProductDetails(iapProduct)) {
-            is IAPProductDetailsResponse.Success -> IAPSupportedResult.Success(isCurrencySupported(response))
-            is IAPProductDetailsResponse.Error -> IAPSupportedResult.Error(response.error)
-        }
-    }
-
     override fun close() {
         iapManager.disconnect()
     }
-
-    private fun isCurrencySupported(response: IAPProductDetailsResponse.Success) =
-        SUPPORTED_CURRENCY.equals(response.productDetails.currencyOfTheFirstPurchasedOffer, ignoreCase = true)
 }
