+++ libs/iap/src/main/java/com/woocommerce/android/iap/pub/PurchaseWPComPlanActions.kt
@@ -1,6 +1,5 @@
 package com.woocommerce.android.iap.pub
 
-import com.woocommerce.android.iap.pub.model.IAPSupportedResult
 import com.woocommerce.android.iap.pub.model.WPComIsPurchasedResult
 import com.woocommerce.android.iap.pub.model.WPComProductResult
 import com.woocommerce.android.iap.pub.model.WPComPurchaseResult
@@ -13,5 +12,4 @@ interface PurchaseWPComPlanActions : Closeable {
     suspend fun isWPComPlanPurchased(): WPComIsPurchasedResult
     suspend fun purchaseWPComPlan(activityWrapper: IAPActivityWrapper)
     suspend fun fetchWPComPlanProduct(): WPComProductResult
-    suspend fun isIAPSupported(): IAPSupportedResult
 }
