+++ libs/iap/src/main/java/com/woocommerce/android/iap/pub/PurchaseWpComPlanSupportChecker.kt
@@ -1,0 +1,8 @@
+package com.woocommerce.android.iap.pub
+
+import com.woocommerce.android.iap.pub.model.IAPSupportedResult
+import java.io.Closeable
+
+interface PurchaseWpComPlanSupportChecker : Closeable {
+    suspend fun isIAPSupported(): IAPSupportedResult
+}
